+++ resources/views/layout/job/input.blade.php
@@ -31,7 +31,7 @@
                 @include($config['blade'].'.input.part')
             </div>
             <div class="tab-pane fade" id="invoice-tab-pane" role="tabpanel" aria-labelledby="invoice-tab" tabindex="0">
-                // invoice
+                @include($config['blade'].'.input.invoice')
             </div>
         @endif
     </div>
@@ -59,6 +59,13 @@
         initJobType(selectJobType, true);
         selectJobType.addEventListener('change', function () {
             initJobType(selectJobType);
+        });
+
+        // select job type
+        let selectJobStatus = document.querySelector('select[name="status"]');
+        initJobStatus(selectJobStatus, true);
+        selectJobStatus.addEventListener('change', function () {
+            initJobStatus(selectJobStatus);
         });
 
         // isi otomatis berdasarkan data warranty konsumen
@@ -285,22 +292,9 @@
 
                 // lanjut jika sukses mengirim data
                 if (obj.status === 'success') {
-                    if ($('input[name="actual_start_at"]').hasAttribute('readonly')) {
-                        $('input[name="actual_start_at"]').removeAttribute('readonly','');
+                    if (!init) {
+                        resetInput();
                     }
-                    if ($('input[name="actual_start_at_time"]').hasAttribute('readonly')) {
-                        $('input[name="actual_start_at_time"]').removeAttribute('readonly','');
-                    }
-                    if ($('input[name="actual_end_at"]').hasAttribute('readonly')) {
-                        $('input[name="actual_end_at"]').removeAttribute('readonly','');
-                    }
-                    if ($('input[name="actual_end_at_time"]').hasAttribute('readonly')) {
-                        $('input[name="actual_end_at_time"]').removeAttribute('readonly','');
-                    }
-                    if ($('input[name="transport"]').hasAttribute('readonly')) {
-                        $('input[name="transport"]').removeAttribute('readonly','');
-                    }
-
                     if (obj.data.actual_date) {
                         $('input[name="actual_start_at"]').setAttribute('readonly','');
                         $('input[name="actual_start_at_time"]').setAttribute('readonly','');
@@ -316,10 +310,179 @@
             }, 'get', null, false)
         }
 
-        let el = $('#table-data-part').querySelector('table tbody');
+        function initJobStatus(el, init = false) {
+            var search = el.options[el.selectedIndex].text;
+            // kirim data untuk menunjukan hasil pencarian
+            let load_url = url('status/?type=form&search='+search);
+            send_http(load_url, function (data) {
+                let obj = JSON.parse(data)
+
+                if (!init) {
+                    //show notification
+                    Toast.fire({
+                        title: ucwords(obj.status),
+                        text: obj.message,
+                        icon: obj.status,
+                        timer: 2000,
+                        timerProgressBar: true,
+                        showConfirmButton: false
+                    })
+                }
+
+                // lanjut jika sukses mengirim data
+                if (obj.status === 'success') {
+                    if (!init) {
+                        resetInput();
+                    }
+                    // jika disable input PARTIAL and FULL
+                    if (obj.data.disable_input > {{ NONE }}) {
+                        $('input[name="flash-fill"]').setAttribute('readonly','');
+                        $('select[name="job_type"]').setAttribute('disabled','');
+                        $('textarea[name="accessories"]').setAttribute('readonly', '');
+                        $('textarea[name="condition"]').setAttribute('readonly', '');
+                        $('textarea[name="service_info"]').setAttribute('readonly', '');
+                        $('textarea[name="repair_info"]').setAttribute('readonly', '');
+                        $('input[name="created_at"]').setAttribute('readonly', '');
+                        $('input[name="created_at_time"]').setAttribute('readonly', '');
+                        $('input[name="repair_at"]').setAttribute('readonly', '');
+                        $('input[name="repair_at_time"]').setAttribute('readonly', '');
+                        $('input[name="collection_at"]').setAttribute('readonly', '');
+                        $('input[name="collection_at_time"]').setAttribute('readonly', '');
+                        $('input[name="actual_start_at"]').setAttribute('readonly', '');
+                        $('input[name="actual_start_at_time"]').setAttribute('readonly', '');
+                        $('input[name="actual_end_at"]').setAttribute('readonly', '');
+                        $('input[name="actual_end_at_time"]').setAttribute('readonly', '');
+
+                        if (obj.data.disable_input > {{ PARTIAL }}) {
+                            $('select[name="status"]').setAttribute('readonly','');
+                            $('input[name="name"]').setAttribute('readonly', '');
+                            $('textarea[name="note"]').setAttribute('readonly', '');
+                            $('input[name="customer_name"]').setAttribute('readonly', '');
+                            $('input[name="phone"]').setAttribute('readonly', '');
+                            $('input[name="phone2"]').setAttribute('readonly', '');
+                            $('input[name="phone3"]').setAttribute('readonly', '');
+                            $('textarea[name="address"]').setAttribute('readonly', '');
+                            $('input[name="email"]').setAttribute('readonly', '');
+                            $('input[name="tax_id"]').setAttribute('readonly', '');
+                            $('select[name="warranty_type"]').setAttribute('disabled','');
+                            $('input[name="model"]').setAttribute('readonly', '');
+                            $('input[name="serial"]').setAttribute('readonly', '');
+                            $('input[name="warranty_no"]').setAttribute('readonly', '');
+                            $('input[name="purchase_date"]').setAttribute('readonly', '');
+                            $('input[name="labour"]').setAttribute('readonly', '');
+                            $('input[name="transport"]').setAttribute('readonly', '');
+                        }
+                    }
+                }
+            }, 'get', null, false)
+        }
+
+        let el = $('#table-data-part');
         if (el) {
-            initTableBtnEdit(el);
-            initTableBtnDelete(el);
+            initTableBtnEdit(el.querySelector('table tbody'));
+            initTableBtnDelete(el.querySelector('table tbody'));
+        }
+
+        function resetInput() {
+            if ($('input[name="flash-fill"]').hasAttribute('readonly')) {
+                $('input[name="flash-fill"]').removeAttribute('readonly','');
+            }
+            if ($('select[name="job_type"]').hasAttribute('disabled')) {
+                $('select[name="job_type"]').removeAttribute('disabled','');
+            }
+            if ($('input[name="name"]').hasAttribute('readonly')) {
+                $('input[name="name"]').removeAttribute('readonly','');
+            }
+            if ($('select[name="status"]').hasAttribute('disabled')) {
+                $('select[name="status"]').removeAttribute('disabled','');
+            }
+            if ($('textarea[name="note"]').hasAttribute('readonly')) {
+                $('textarea[name="note"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="customer_name"]').hasAttribute('readonly')) {
+                $('input[name="customer_name"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="phone"]').hasAttribute('readonly')) {
+                $('input[name="phone"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="phone2"]').hasAttribute('readonly')) {
+                $('input[name="phone2"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="phone3"]').hasAttribute('readonly')) {
+                $('input[name="phone3"]').removeAttribute('readonly','');
+            }
+            if ($('textarea[name="address"]').hasAttribute('readonly')) {
+                $('textarea[name="address"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="email"]').hasAttribute('readonly')) {
+                $('input[name="email"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="tax_id"]').hasAttribute('readonly')) {
+                $('input[name="tax_id"]').removeAttribute('readonly','');
+            }
+            if ($('select[name="warranty_type"]').hasAttribute('disabled')) {
+                $('select[name="warranty_type"]').removeAttribute('disabled','');
+            }
+            if ($('input[name="model"]').hasAttribute('readonly')) {
+                $('input[name="model"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="serial"]').hasAttribute('readonly')) {
+                $('input[name="serial"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="warranty_no"]').hasAttribute('readonly')) {
+                $('input[name="warranty_no"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="purchase_date"]').hasAttribute('readonly')) {
+                $('input[name="purchase_date"]').removeAttribute('readonly','');
+            }
+            if ($('textarea[name="accessories"]').hasAttribute('readonly')) {
+                $('textarea[name="accessories"]').removeAttribute('readonly','');
+            }
+            if ($('textarea[name="condition"]').hasAttribute('readonly')) {
+                $('textarea[name="condition"]').removeAttribute('readonly','');
+            }
+            if ($('textarea[name="service_info"]').hasAttribute('readonly')) {
+                $('textarea[name="service_info"]').removeAttribute('readonly','');
+            }
+            if ($('textarea[name="repair_info"]').hasAttribute('readonly')) {
+                $('textarea[name="repair_info"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="created_at"]').hasAttribute('readonly')) {
+                $('input[name="created_at"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="created_at_time"]').hasAttribute('readonly')) {
+                $('input[name="created_at_time"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="repair_at"]').hasAttribute('readonly')) {
+                $('input[name="repair_at"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="repair_at_time"]').hasAttribute('readonly')) {
+                $('input[name="repair_at_time"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="collection_at"]').hasAttribute('readonly')) {
+                $('input[name="collection_at"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="collection_at_time"]').hasAttribute('readonly')) {
+                $('input[name="collection_at_time"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="actual_start_at"]').hasAttribute('readonly')) {
+                $('input[name="actual_start_at"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="actual_start_at_time"]').hasAttribute('readonly')) {
+                $('input[name="actual_start_at_time"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="actual_end_at"]').hasAttribute('readonly')) {
+                $('input[name="actual_end_at"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="actual_end_at_time"]').hasAttribute('readonly')) {
+                $('input[name="actual_end_at_time"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="labour"]').hasAttribute('readonly')) {
+                $('input[name="labour"]').removeAttribute('readonly','');
+            }
+            if ($('input[name="transport"]').hasAttribute('readonly')) {
+                $('input[name="transport"]').removeAttribute('readonly','');
+            }
         }
     </script>
 @endsection